refactor files in their correct folders
