+++ src/main/java/BusinessLayer/Helper/ParseLocalDateTime.java
@@ -1,8 +1,8 @@
-package Helper;
+package BusinessLayer.Helper;
 
 import java.time.LocalDateTime;
 
-public class parseLocalDateTime {
+public class ParseLocalDateTime {
     // example dateTimeStr:
     // 2 0 2 1 - 0 7 - 1 1 T  1  6  :  4  4  :  0  0 +07:00
     // 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18
