+++ src/main/java/BusinessLayer/Helper/Scraper.java
@@ -1,13 +1,8 @@
-package Scraper;
+package BusinessLayer.Helper;
 
 import org.jsoup.*;
 import org.jsoup.nodes.*;
 import org.jsoup.select.*;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 
 import java.io.*;
 import java.net.*;
