+++ src/main/java/BusinessLayer/News/Preview.java
@@ -1,4 +1,4 @@
-package News;
+package BusinessLayer.News;
 
 import org.jsoup.nodes.Element;
