+++ src/main/java/BusinessLayer/NewsSources/TuoiTre.java
@@ -1,8 +1,10 @@
-package News;
+package BusinessLayer.NewsSources;
 
-import News.Sanitizer.HtmlSanitizer;
-import News.Sanitizer.TuoiTreSanitizer;
-import News.Sanitizer.ZingNewsSanitizer;
+import BusinessLayer.Helper.CATEGORY;
+import BusinessLayer.Helper.CSS;
+import BusinessLayer.Helper.ParseLocalDateTime;
+import BusinessLayer.Sanitizer.HtmlSanitizer;
+import BusinessLayer.Sanitizer.TuoiTreSanitizer;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
@@ -57,7 +59,7 @@ public class TuoiTre extends NewsOutlet{
         if (dateTimeStr.isEmpty() || dateTimeStr.isBlank()){
             return LocalDateTime.now();
         }
-        return Helper.parseLocalDateTime.parse(dateTimeStr);
+        return ParseLocalDateTime.parse(dateTimeStr);
     }
 
     @Override
