+++ src/main/java/BusinessLayer/NewsSources/VNExpress.java
@@ -1,7 +1,10 @@
-package News;
+package BusinessLayer.NewsSources;
 
-import News.Sanitizer.HtmlSanitizer;
-import News.Sanitizer.VNExpressSanitizer;
+import BusinessLayer.Helper.CATEGORY;
+import BusinessLayer.Helper.CSS;
+import BusinessLayer.Helper.ParseLocalDateTime;
+import BusinessLayer.Sanitizer.HtmlSanitizer;
+import BusinessLayer.Sanitizer.VNExpressSanitizer;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
@@ -56,7 +59,7 @@ public class VNExpress extends NewsOutlet{
         if (dateTimeStr.isEmpty() || dateTimeStr.isBlank()){
             return LocalDateTime.now();
         }
-        return Helper.parseLocalDateTime.parse(dateTimeStr);
+        return ParseLocalDateTime.parse(dateTimeStr);
     }
 
     @Override
