+++ src/main/java/BusinessLayer/Sanitizer/NhanDanSanitizer.java
@@ -1,14 +1,12 @@
-package News.Sanitizer;
+package BusinessLayer.Sanitizer;
 
-import News.CSS;
+import BusinessLayer.Helper.CSS;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Element;
 import org.jsoup.nodes.Node;
 import org.jsoup.safety.Safelist;
 import org.jsoup.select.NodeFilter;
 import org.jsoup.select.NodeTraversor;
-
-import java.util.logging.Filter;
 
 public class NhanDanSanitizer extends HtmlSanitizer {
     @Override
